pyproject.toml
[project]
name = "azure-container-apps-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-mgmt-appcontainers>=4.0.0",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.2",
]
